requirements.txt
ipykernel==6.29.3
ipython==8.18.1
jupyter_client==8.6.1
jupyter_core==5.7.2
matplotlib==3.8.3
numpy==1.26.4
scikit-learn==1.4.1.post1
scipy==1.12.0
torch==2.2.0
tqdm==4.66.2
pillow==10.3.0
torchvision==0.17.1